% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiodata-package.R
\docType{package}
\name{fiodata-package}
\alias{fiodata}
\alias{fiodata-package}
\title{fiodata: Regional and Multi-Regional Input-Output Data}
\description{
Provides Regional (Brazil, 2020) and Multi-Regional (World, 2000) input-output matrices for R. This package serves as a data-only companion to the 'fio' package, facilitating input-output analysis by providing standardized R6 data objects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/albersonmiranda/fiodata}
  \item Report bugs at \url{https://github.com/albersonmiranda/fiodata/issues}
}

}
\author{
\strong{Maintainer}: Alberson da Silva Miranda \email{albersonmiranda@hotmail.com} (\href{https://orcid.org/0000-0001-9252-4175}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Celso Bissoli Sessa \email{celso.sessa@ufes.br} (\href{https://orcid.org/0000-0001-7616-0244}{ORCID}) [data contributor]
}

}
\keyword{internal}
